chore: move _utilities under internal
We move the `_utilities` package under `internal`, as the functions are
not external facing. We also use the init file in `_utilities` to roll
up the names from the data validation module.
