+++ src/momento/internal/common/_data_client_ops.py
@@ -13,13 +13,9 @@ from momento_wire_types.cacheclient_pb2 import (
 )
 
 from momento import logs
-from momento._utilities._data_validation import (
-    _as_bytes,
-    _validate_cache_name,
-    _validate_ttl,
-)
 from momento.config.configuration import Configuration
 from momento.errors import SdkException, convert_error
+from momento.internal._utilities import _as_bytes, _validate_cache_name, _validate_ttl
 from momento.responses import (
     CacheDelete,
     CacheDeleteResponse,
