+++ src/momento/internal/synchronous/_scs_control_client.py
@@ -13,9 +13,9 @@ from momento_wire_types.controlclient_pb2 import (
 from momento_wire_types.controlclient_pb2_grpc import ScsControlStub
 
 from momento import logs
-from momento._utilities._data_validation import _validate_cache_name, _validate_ttl
 from momento.auth.credential_provider import CredentialProvider
 from momento.errors import convert_error
+from momento.internal._utilities import _validate_cache_name, _validate_ttl
 from momento.internal.synchronous._scs_grpc_manager import _ControlGrpcManager
 from momento.responses import (
     CreateCache,
